fix(package.json) Move AngularJS as a `peerDependency`
AngularJS is now a `peerDependency` and it has to be installed separately. This is a breaking change, so the `package.json`
version was bumped to a `1.0.0`. The default NPM dependency resolution usually works just fine with the explicit dependency, but it
also introduce some weird behaviors, like allowing two versions of the same library to co-exist. Since it is generally a _bad idea_ to have
two concurrent AngularJS versions installed, turning it into a `peerDependency` makes more sense and give the user of this library all
control over AngularJS installation.